backend-api/requirements.txt
# API Framework
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3

# Database
psycopg2-binary>=2.9.9
tenacity>=8.2.3

# AI Services
anthropic>=0.18.0
openai>=1.1.1
google-generativeai>=0.7.0

# Environment and Utils
python-dotenv>=1.0.0
python-multipart>=0.0.6
pytest>=7.4.3
httpx>=0.25.0